Add files via upload
Sentiment analysis is an application of natural language processing. It is also known as emotion extraction
or opinion mining. This is a very popular field of research in text mining. Numerous social media sites have
enabled the exchange of opinions among users all over the world which has promoted the use of popular
social network sites such as Twitter, for communication. The users’ tweets are highly unstructured,
heterogeneous, and vulgar, and they cover a wide range of topics.